requirements.txt
# conda install -c skmad simpleaudio
# conda install scikit-image 
# conda install -c huggingface transformers
# conda install -c intel scikit-learn 
# conda install -c anaconda cython  
# pip install phonemizer   
# conda install -c conda-forge unidecode 
# conda install scipy  
# pip install wavio  
# pip install opencv-python   

# cd monotonic_align; python setup.py build_ext --inplace
# cd ..
________________
# pip install flask
# pip install Flask-SQLAlchemy
# pip install flask-login

attrs==21.4.0
boto3==1.16.0
botocore==1.19.1
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clip @ git+https://github.com/openai/CLIP.git@40f5484c1c74edd83cb9cf687c6ab92b28d8b656
clldutils==3.10.1
cloudpickle==1.6.0
colorlog==6.6.0
cryptography==3.1.1
csvw==2.0.0
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask==2.30.0
dataclasses==0.7
decorator==4.4.2
dlinfo==1.2.1
docutils==0.15.2
filelock==3.0.12
ftfy==6.0.3
huggingface-hub==0.0.17
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
isodate==0.6.1
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.1
networkx==2.5
numpy==1.19.1
olefile==0.46
opencv-python==4.5.5.62
packaging==20.4
phonemizer==3.0.1
Pillow==8.0.0
pip==20.2.4
protobuf==3.13.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2022.1.18
requests==2.24.0
rfc3986==1.5.0
s3transfer==0.3.3
sacremoses==0.0.46
scikit-image==0.17.2
scikit-learn==0.20.0
scipy==1.5.2
segments==2.2.0
setuptools==50.3.0.post20201006
simpleaudio==1.0.2
six==1.15.0
tabulate==0.8.9
tokenizers==0.10.3
toolz==0.11.1
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tornado==6.0.4
tqdm==4.50.2
transformers==4.11.3
typing-extensions==3.7.4.3
Unidecode==1.3.3
uritemplate==4.1.1
urllib3==1.25.11
wavio==0.0.4
wcwidth==0.2.5
wheel==0.35.1
zipp==3.3.1
